# server/target/openapi.yaml
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Control Finanzas API",
                 "description":  "API para cuentas y ahorro (iteraciÃ³n inicial)",
                 "version":  "0.1.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "paths":  {
...
                  "/api/v1/savings/plans":  {
                                                "get":  {
                                                            "tags":  [
                                                                         "savings-controller"
                                                                     ],
                                                            "operationId":  "listPlans",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "arg0",
                                                                                   "in":  "query",
                                                                                   "required":  false,
                                                                                   "schema":  {
                                                                                                  "type":  "string"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "OK",
                                                                                          "content":  {
                                                                                                          "*/*":  {
                                                                                                                      "schema":  {
                                                                                                                                     "type":  "array",
                                                                                                                                     "items":  {
                                                                                                                                                   "$ref":  "#/components/schemas/PlanView"
                                                                                                                                               }
                                                                                                                                 }
                                                                                                                  }
                                                                                                      }
                                                                                      }
                                                                          }
                                                        },
                                                "post":  {
                                                             "tags":  [
                                                                          "savings-controller"
                                                                      ],
                                                             "operationId":  "createPlan",
                                                             "requestBody":  {
                                                                                 "content":  {
                                                                                                 "application/json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/PlanCreate"
                                                                                                                                     }
                                                                                                                      }
                                                                                             },
                                                                                 "required":  true
                                                                             },
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "type":  "object"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       }
                                                                           }
                                                         }
                                            },
...
                  "/api/v1/savings/moves/schedule":  {
                                                         "post":  {
                                                                      "tags":  [
                                                                                   "savings-controller"
                                                                               ],
                                                                      "operationId":  "schedule",
                                                                      "parameters":  [
                                                                                         {
                                                                                             "name":  "arg0",
                                                                                             "in":  "query",
                                                                                             "required":  true,
                                                                                             "schema":  {
                                                                                                            "type":  "string"
                                                                                                        }
                                                                                         },
                                                                                         {
                                                                                             "name":  "arg1",
                                                                                             "in":  "query",
                                                                                             "required":  false,
                                                                                             "schema":  {
                                                                                                            "type":  "number",
                                                                                                            "format":  "double"
                                                                                                        }
                                                                                         }
                                                                                     ],
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "OK",
                                                                                                    "content":  {
                                                                                                                    "*/*":  {
                                                                                                                                "schema":  {
                                                                                                                                               "type":  "array",
                                                                                                                                               "items":  {
                                                                                                                                                             "$ref":  "#/components/schemas/MoveView"
                                                                                                                                                         }
                                                                                                                                           }
                                                                                                                            }
                                                                                                                }
                                                                                                }
                                                                                    }
                                                                  }
                                                     },
...
                  "/api/v1/savings/plans/{id}":  {
                                                     "delete":  {
                                                                    "tags":  [
                                                                                 "savings-controller"
                                                                             ],
                                                                    "operationId":  "deletePlan",
                                                                    "parameters":  [

                                                                                   ],
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "OK",
                                                                                                  "content":  {
                                                                                                                  "*/*":  {
                                                                                                                              "schema":  {
                                                                                                                                             "type":  "object"
                                                                                                                                         }
                                                                                                                          }
                                                                                                              }
                                                                                              }
                                                                                  }
                                                                },
                                                     "patch":  {
                                                                   "tags":  [
                                                                                "savings-controller"
                                                                            ],
                                                                   "operationId":  "updatePlan",
                                                                   "parameters":  [

                                                                                  ],
                                                                   "requestBody":  {
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/PlanUpdate"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   },
                                                                                       "required":  true
                                                                                   },
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "OK",
                                                                                                 "content":  {
                                                                                                                 "*/*":  {
                                                                                                                             "schema":  {
                                                                                                                                            "type":  "object"
                                                                                                                                        }
                                                                                                                         }
                                                                                                             }
                                                                                             }
                                                                                 }
                                                               }
                                                 },
...
                  "/api/v1/savings/moves":  {
                                                "get":  {
                                                            "tags":  [
                                                                         "savings-controller"
                                                                     ],
                                                            "operationId":  "listMoves",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "arg0",
                                                                                   "in":  "query",
                                                                                   "required":  false,
                                                                                   "schema":  {
                                                                                                  "type":  "string"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "OK",
                                                                                          "content":  {
                                                                                                          "*/*":  {
                                                                                                                      "schema":  {
                                                                                                                                     "type":  "array",
                                                                                                                                     "items":  {
                                                                                                                                                   "$ref":  "#/components/schemas/MoveView"
                                                                                                                                               }
                                                                                                                                 }
                                                                                                                  }
                                                                                                      }
                                                                                      }
                                                                          }
                                                        }
                                            },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "PlanCreate":  {
                                                          "required":  [
                                                                           "amountPlanned",
                                                                           "name",
                                                                           "type"
                                                                       ],
                                                          "type":  "object",
                                                          "properties":  {
                                                                             "name":  {
                                                                                          "maxLength":  100,
                                                                                          "minLength":  0,
                                                                                          "type":  "string"
                                                                                      },
                                                                             "monthKey":  {
                                                                                              "pattern":  "^[0-9]{4}-[0-9]{2}$",
                                                                                              "type":  "string"
                                                                                          },
                                                                             "type":  {
                                                                                          "type":  "string",
                                                                                          "enum":  [
                                                                                                       "FIXED",
                                                                                                       "PERCENT"
                                                                                                   ]
                                                                                      },
                                                                             "amountPlanned":  {
                                                                                                   "type":  "number",
                                                                                                   "format":  "double"
                                                                                               },
                                                                             "percent":  {
                                                                                             "maximum":  1.0,
                                                                                             "exclusiveMaximum":  false,
                                                                                             "minimum":  0.0,
                                                                                             "exclusiveMinimum":  false,
                                                                                             "type":  "number",
                                                                                             "format":  "double"
                                                                                         },
                                                                             "priority":  {
                                                                                              "minimum":  1,
                                                                                              "type":  "integer",
                                                                                              "format":  "int32"
                                                                                          },
                                                                             "status":  {
                                                                                            "type":  "string",
                                                                                            "enum":  [
                                                                                                         "ACTIVE",
                                                                                                         "PAUSED"
                                                                                                     ]
                                                                                        },
                                                                             "sourceAccountId":  {
                                                                                                     "type":  "integer",
                                                                                                     "format":  "int64"
                                                                                                 },
                                                                             "targetAccountId":  {
                                                                                                     "type":  "integer",
                                                                                                     "format":  "int64"
                                                                                                 }
                                                                         }
                                                      },
...
                                       "PlanView":  {
                                                        "type":  "object",
                                                        "properties":  {
                                                                           "id":  {
                                                                                      "type":  "integer",
                                                                                      "format":  "int64"
                                                                                  },
                                                                           "name":  {
                                                                                        "type":  "string"
                                                                                    },
                                                                           "monthKey":  {
                                                                                            "type":  "string"
                                                                                        },
                                                                           "type":  {
                                                                                        "type":  "string",
                                                                                        "enum":  [
                                                                                                     "FIXED",
                                                                                                     "PERCENT"
                                                                                                 ]
                                                                                    },
                                                                           "amountPlanned":  {
                                                                                                 "type":  "number",
                                                                                                 "format":  "double"
                                                                                             },
                                                                           "percent":  {
                                                                                           "type":  "number",
                                                                                           "format":  "double"
                                                                                       },
                                                                           "priority":  {
                                                                                            "type":  "integer",
                                                                                            "format":  "int32"
                                                                                        },
                                                                           "status":  {
                                                                                          "type":  "string",
                                                                                          "enum":  [
                                                                                                       "ACTIVE",
                                                                                                       "PAUSED"
                                                                                                   ]
                                                                                      },
                                                                           "sourceAccountId":  {
                                                                                                   "type":  "integer",
                                                                                                   "format":  "int64"
                                                                                               },
                                                                           "targetAccountId":  {
                                                                                                   "type":  "integer",
                                                                                                   "format":  "int64"
                                                                                               }
                                                                       }
                                                    },
...
                                   }
                   }
}